Azure/requirements.txt
pyodbc
python-dotenv
